Share optimizing code between fsharp and minimal
