+++ packages/babel-plugin-proposal-pipeline-operator/test/fixtures/minimal/arrow-functions/output.js
@@ -1,6 +1,6 @@
-var _sum, _ref, _ref2;
+var _ref, _ref2, _ref3;
 
-var result = (_sum = (_ref = (_ref2 = [5, 10], _ref2.map(x => x * 2)), _ref.reduce((a, b) => a + b)), _sum + 1);
+var result = (_ref = (_ref2 = (_ref3 = [5, 10], _ref3.map(x => x * 2)), _ref2.reduce((a, b) => a + b)), _ref + 1);
 expect(result).toBe(31);
 
 var inc = x => x + 1;
@@ -8,8 +8,8 @@ var inc = x => x + 1;
 var double = x => x * 2;
 
 var result2 = [4, 9].map(x => {
-  var _ref3, _x;
+  var _ref4, _x;
 
-  return _ref3 = (_x = x, inc(_x)), double(_ref3);
+  return _ref4 = (_x = x, inc(_x)), double(_ref4);
 });
 expect(result2).toEqual([10, 20]);
