+++ packages/babel-plugin-proposal-pipeline-operator/test/fixtures/minimal/optimize-0-param-arrow/output.js
@@ -1,3 +1,5 @@
+var _ref, _a;
+
 var a = 1,
     b = 2,
     c = 3;
